requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.6.1
paho-mqtt==1.5.0
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
Werkzeug==1.0.1
